+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/intellij/documentation/XQueryDocumentationProvider.kt
@@ -17,7 +17,9 @@ package uk.co.reecedunn.intellij.plugin.intellij.documentation
 
 import com.intellij.lang.documentation.AbstractDocumentationProvider
 import com.intellij.psi.PsiElement
+import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathNCName
 import uk.co.reecedunn.intellij.plugin.xpath.ast.xpath.XPathVarName
+import uk.co.reecedunn.intellij.plugin.xpath.model.XPathNamespaceDeclaration
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryFunctionDecl
 import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryVarDecl
 
@@ -34,6 +36,13 @@ object XQueryDocumentationProvider : AbstractDocumentationProvider() {
                 (parent.parent as? XQueryVarDecl)?.let {
                     val sig = it.presentation?.presentableText
                     "declare variable $sig"
+                }
+            }
+            is XPathNCName -> {
+                (parent.parent as? XPathNamespaceDeclaration)?.let { decl ->
+                    val prefix = decl.namespacePrefix?.data
+                    val uri = decl.namespaceUri?.data ?: return null
+                    prefix?.let { "namespace $it = \"$uri\"" } ?: "namespace \"{$uri}\""
                 }
             }
             else -> null