+++ src/lang-xquery/test/uk/co/reecedunn/intellij/plugin/intellij/tests/documentation/XQueryDocumentationProviderTest.kt
@@ -17,6 +17,7 @@ package uk.co.reecedunn.intellij.plugin.intellij.tests.documentation
 
 import com.intellij.psi.PsiElement
 import org.hamcrest.CoreMatchers.`is`
+import org.hamcrest.CoreMatchers.nullValue
 import org.junit.jupiter.api.*
 import uk.co.reecedunn.intellij.plugin.core.sequences.walkTree
 import uk.co.reecedunn.intellij.plugin.core.tests.assertion.assertThat
@@ -31,6 +32,99 @@ import uk.co.reecedunn.intellij.plugin.xquery.tests.parser.ParserTestCase
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @DisplayName("IntelliJ - Custom Language Support - Documentation Provider - XQuery")
 private class XQueryDocumentationProviderTest : ParserTestCase() {
+    @Nested
+    @DisplayName("XQuery 3.1 EBNF (23) ModuleImport")
+    internal inner class ModuleImport {
+        fun parse(text: String): Pair<PsiElement?, PsiElement?> {
+            val module = parseText(text)
+            val call = module.walkTree().filterIsInstance<XPathFunctionCall>().first() as XPathFunctionReference
+            val element = call.functionName?.element!!
+            val ref = element.references[0].resolve()
+            return element to ref
+        }
+
+        @Test
+        @DisplayName("custom")
+        fun custom() {
+            val ref = parse(
+                """
+                import module namespace ex = "http://www.example.com";
+                declare function ex:test((::)) {};
+                ex:test()
+                """
+            )
+
+            val quickDoc = XQueryDocumentationProvider.getQuickNavigateInfo(ref.second, ref.first)
+            assertThat(quickDoc, `is`("namespace ex = \"http://www.example.com\""))
+        }
+
+        @Test
+        @DisplayName("custom without namespace uri")
+        fun customWithoutUri() {
+            val ref = parse(
+                """
+                import module namespace ex;
+                declare function ex:test((::)) {};
+                ex:test()
+                """
+            )
+
+            val quickDoc = XQueryDocumentationProvider.getQuickNavigateInfo(ref.second, ref.first)
+            assertThat(quickDoc, `is`(nullValue()))
+        }
+    }
+
+    @Nested
+    @DisplayName("XQuery 3.1 EBNF (24) NamespaceDecl")
+    internal inner class NamespaceDecl {
+        fun parse(text: String): Pair<PsiElement?, PsiElement?> {
+            val module = parseText(text)
+            val call = module.walkTree().filterIsInstance<XPathFunctionCall>().first() as XPathFunctionReference
+            val element = call.functionName?.element!!
+            val ref = element.references[0].resolve()
+            return element to ref
+        }
+
+        @Test
+        @DisplayName("builtin")
+        fun builtin() {
+            val ref = parse("fn:true()")
+
+            val quickDoc = XQueryDocumentationProvider.getQuickNavigateInfo(ref.second, ref.first)
+            assertThat(quickDoc, `is`("namespace fn = \"http://www.w3.org/2005/xpath-functions\""))
+        }
+
+        @Test
+        @DisplayName("custom")
+        fun custom() {
+            val ref = parse(
+                """
+                declare namespace ex = "http://www.example.com";
+                declare function ex:test((::)) {};
+                ex:test()
+                """
+            )
+
+            val quickDoc = XQueryDocumentationProvider.getQuickNavigateInfo(ref.second, ref.first)
+            assertThat(quickDoc, `is`("namespace ex = \"http://www.example.com\""))
+        }
+
+        @Test
+        @DisplayName("custom without namespace uri")
+        fun customWithoutUri() {
+            val ref = parse(
+                """
+                declare namespace ex;
+                declare function ex:test((::)) {};
+                ex:test()
+                """
+            )
+
+            val quickDoc = XQueryDocumentationProvider.getQuickNavigateInfo(ref.second, ref.first)
+            assertThat(quickDoc, `is`(nullValue()))
+        }
+    }
+
     @Nested
     @DisplayName("XQuery 3.1 EBNF (131) VarRef")
     internal inner class VarRef {